requirements.txt
kafka-python==2.0.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
schedule==1.2.1
